Create risk vector based on ratio, number of places, or data frame

Create risk vector based on ratio, number of places, or data frame

rv_create_risk_vector(
  risk_ratio,
  total_places,
  risk_places = NULL,
  risk_name = "Risk"
)

Arguments

risk_ratio
A numeic vector or a data frame with two columns, `Risk` and `Ratio`

total_places
Numeric. Gives the total length of the output vector.

risk_places
A numeric vector or a data frame with two columns, `Risk` and `Places`

risk_name
If a numeric vector given in previous parameters, defines the name of the risk.

Value

A character vector with length equal to total_places.

Examples

rv_create_risk_vector(risk_ratio = 0.1,
                      total_places = 100)

#>   [1] NA       NA       NA       NA       NA       NA       "Risk 1" "Risk 1"
#>   [9] NA       NA       NA       NA       NA       NA       NA       NA      
#>  [17] NA       NA       NA       NA       NA       NA       NA       NA      
#>  [25] NA       NA       "Risk 1" NA       "Risk 1" NA       NA       NA      
#>  [33] NA       NA       NA       NA       NA       NA       NA       "Risk 1"
#>  [41] NA       NA       NA       NA       NA       NA       NA       NA      
#>  [49] NA       NA       "Risk 1" "Risk 1" NA       NA       NA       NA      
#>  [57] NA       NA       NA       NA       NA       NA       NA       NA      
#>  [65] NA       NA       NA       NA       NA       NA       NA       NA      
#>  [73] NA       NA       NA       NA       NA       NA       NA       NA      
#>  [81] NA       NA       NA       NA       NA       "Risk 1" NA       NA      
#>  [89] NA       "Risk 1" NA       NA       NA       NA       NA       NA      
#>  [97] "Risk 1" NA       NA       NA      

table(
  rv_create_risk_vector(risk_ratio = 0.1,
                        total_places = 100) 
)

#> 
#> Risk 1 
#>     10 

ratio <- tibble::tribble(~Risk, ~Ratio,
                         "Hospitalization", 0.3,
                         "Death", 0.1)

rv_create_risk_vector(risk_ratio = ratio,
                      total_places = 100)

#>   [1] NA                NA                "Death"           NA               
#>   [5] NA                NA                "Hospitalization" "Hospitalization"
#>   [9] "Hospitalization" NA                NA                NA               
#>  [13] NA                NA                NA                NA               
#>  [17] NA                "Death"           NA                "Death"          
#>  [21] NA                NA                NA                "Hospitalization"
#>  [25] NA                NA                "Death"           NA               
#>  [29] NA                NA                NA                NA               
#>  [33] "Hospitalization" NA                NA                NA               
#>  [37] NA                NA                "Hospitalization" NA               
#>  [41] "Hospitalization" NA                "Hospitalization" NA               
#>  [45] NA                "Death"           "Hospitalization" NA               
#>  [49] NA                "Hospitalization" "Hospitalization" NA               
#>  [53] "Hospitalization" NA                NA                NA               
#>  [57] "Death"           NA                "Hospitalization" NA               
#>  [61] "Hospitalization" NA                NA                "Death"          
#>  [65] "Hospitalization" "Hospitalization" NA                NA               
#>  [69] "Hospitalization" "Hospitalization" NA                NA               
#>  [73] "Hospitalization" NA                "Hospitalization" "Hospitalization"
#>  [77] NA                NA                NA                "Hospitalization"
#>  [81] NA                "Death"           NA                "Hospitalization"
#>  [85] NA                "Hospitalization" "Hospitalization" NA               
#>  [89] "Hospitalization" NA                "Hospitalization" NA               
#>  [93] "Hospitalization" "Death"           NA                "Death"          
#>  [97] NA                "Hospitalization" "Hospitalization" NA               

table(
  rv_create_risk_vector(risk_ratio = ratio,
                        total_places = 100)
)

#> 
#>           Death Hospitalization 
#>              10              30